+++ bedrock/bedrock-v388/src/main/java/com/nukkitx/protocol/bedrock/v388/serializer/AvailableCommandsSerializer_v388.java
@@ -107,7 +107,7 @@ public class AvailableCommandsSerializer_v388 extends AvailableCommandsSerialize
 
         // Generate command data
         for (CommandData.Builder command : commands) {
-            byte flags = command.getFlags();
+            int flags = command.getFlags();
             List<CommandData.Flag> flagList = new ObjectArrayList<>();
             for (int i = 0; i < 6; i++) {
                 if ((flags & (1 << i)) != 0) {
